Q-A_Retrieval/requirements.txt
accelerate==0.28.0
datasets==2.18.0
evaluate==0.4.1
tqdm
einops
sentencepiece
scikit-learn
matplotlib
pandas
numpy
transformers==4.39.1
seaborn
sentence-transformers
flask
flask-cors
ollama
pymupdf
python-docx
